Media 완료 흐름을 resolver에 드러낸다
